+++ src/Sulu/Component/Media/SystemCollections/SystemCollectionBuilder.php
@@ -49,7 +49,7 @@ class SystemCollectionBuilder implements BuilderInterface, ContainerAwareInterfa
      */
     public function getDependencies()
     {
-        return ['database', 'fixtures'];
+        return ['cache', 'database', 'fixtures'];
     }
 
     /**